requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
urllib3==1.25.10
Werkzeug==1.0.1
